+++ libs/nx-plugin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.47.4](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.10...v0.47.4) (2023-11-28)
+
+**Note:** Version bump only for package @fundamental-ngx/nx-plugin
+
+
+
+
+
+## [0.47.4-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.9...v0.47.4-rc.10) (2023-11-28)
+
+**Note:** Version bump only for package @fundamental-ngx/nx-plugin
+
+
+
+
+
+## [0.47.4-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.8...v0.47.4-rc.9) (2023-11-28)
+
+**Note:** Version bump only for package @fundamental-ngx/nx-plugin
+
+
+
+
+
 ## [0.47.4-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.47.4-rc.7...v0.47.4-rc.8) (2023-11-27)
 
 **Note:** Version bump only for package @fundamental-ngx/nx-plugin